Refactor `Parser` into `Loader`
* `Parser.parse` -> `Loader.load`
* `Parser._parse -> `Loader._load`
* Use `pytz.open_resource` for zone file loading to fix pytz installs
  using shared zoneinfo files.
* Fix for python2 setup and test (backport `FileNotFoundError` name).

Tested on Fedora Core 24 (linux), python2.7 (distro), python3.5
(distro)

Fixes #34
