--- pendulum/tz/parser.py
@@ -1,123 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from struct import unpack, calcsize
-from datetime import datetime
-
-from .transition import Transition
-from .transition_type import TransitionType
-from .breakdown import local_time
-
-
-def _byte_string(s):
-    """Cast a string or byte string to an ASCII byte string."""
-    return s.encode('US-ASCII')
-
-_NULL = _byte_string('\0')
-
-
-def _std_string(s):
-    """Cast a string or byte string to an ASCII string."""
-    return str(s.decode('US-ASCII'))
-
-
-class Parser(object):
-
-    @classmethod
-    def parse(cls, filepath):
-        with open(filepath, 'rb') as fp:
-            return cls._parse(fp)
-
-    @classmethod
-    def _parse(cls, fp):
-        head_fmt = '>4s c 15x 6l'
-        head_size = calcsize(head_fmt)
-        (magic, fmt, ttisgmtcnt, ttisstdcnt, leapcnt, timecnt,
-         typecnt, charcnt) = unpack(head_fmt, fp.read(head_size))
-
-        # Make sure it is a tzfile(5) file
-        assert magic == _byte_string('TZif'), 'Got magic %s' % repr(magic)
-
-        # Read out the transition times,
-        # localtime indices and ttinfo structures.
-        data_fmt = '>%(timecnt)dl %(timecnt)dB %(ttinfo)s %(charcnt)ds' % dict(
-            timecnt=timecnt, ttinfo='lBB' * typecnt, charcnt=charcnt)
-        data_size = calcsize(data_fmt)
-        data = unpack(data_fmt, fp.read(data_size))
-
-        # make sure we unpacked the right number of values
-        assert len(data) == 2 * timecnt + 3 * typecnt + 1
-        transition_times = tuple(trans for trans in data[:timecnt])
-        lindexes = tuple(data[timecnt:2 * timecnt])
-        ttinfo_raw = data[2 * timecnt:-1]
-        tznames_raw = data[-1]
-        del data
-
-        # Process ttinfo into separate structs
-        transition_types = tuple()
-        tznames = {}
-        i = 0
-        while i < len(ttinfo_raw):
-            # have we looked up this timezone name yet?
-            tzname_offset = ttinfo_raw[i + 2]
-            if tzname_offset not in tznames:
-                nul = tznames_raw.find(_NULL, tzname_offset)
-                if nul < 0:
-                    nul = len(tznames_raw)
-                tznames[tzname_offset] = _std_string(
-                    tznames_raw[tzname_offset:nul])
-            transition_types += (
-                TransitionType(
-                    ttinfo_raw[i], bool(ttinfo_raw[i + 1]),
-                    tznames[tzname_offset]
-                ),
-            )
-            i += 3
-
-        # Now build the timezone object
-        if len(transition_times) == 0:
-            transitions = tuple()
-        else:
-            # calculate transition info
-            transitions = tuple()
-            for i in range(len(transition_times)):
-                transition_type = transition_types[lindexes[i]]
-
-                if i == 0:
-                    pre_transition_type = transition_types[lindexes[i]]
-                else:
-                    pre_transition_type = transition_types[lindexes[i - 1]]
-
-                pre_time = datetime(*local_time(transition_times[i],
-                                                pre_transition_type)[:7])
-                time = datetime(*local_time(transition_times[i],
-                                            transition_type)[:7])
-                tr = Transition(
-                    transition_times[i],
-                    transition_type,
-                    pre_time,
-                    time,
-                    pre_transition_type
-                )
-
-                transitions += (tr,)
-
-        # Determine the before-first-transition type
-        default_transition_type_index = 0
-        if transitions:
-            index = 0
-            if transition_types[0].is_dst:
-                index = transition_types.index(transitions[0].transition_type)
-                while index != 0 and transition_types[index].is_dst:
-                    index -= 1
-
-            while index != len(transitions) and transition_types[index].is_dst:
-                index += 1
-
-            if index != len(transitions):
-                default_transition_type_index = index
-
-        return (
-            transitions,
-            transition_types,
-            transition_types[default_transition_type_index]
-        )
